CMake rewrite

+++ doc/sphinxman/document_tests.pl
@@ -29,6 +29,8 @@ my %ExeFolder = (
    "libefp/"    => "libefp",
    "pcmsolver/" => "pcmsolver",
    "dmrg/"      => "dmrg",
+   "gdma/"      => "gdma",
+   "dkh/"       => "dkh",
 );
 
 foreach my $exe (keys %ExeFolder) {
@@ -48,6 +50,8 @@ foreach my $File(readdir SAMPLES){
     next if $File =~ /^libefp$/;
     next if $File =~ /^pcmsolver$/;
     next if $File =~ /^dmrg$/;
+    next if $File =~ /^gdma$/;
+    next if $File =~ /^dkh$/;
     next if (-d $File);  # Don't remove subdirectories
     remove_tree("$SamplesFolder/$File");
 }
